requirements.txt
streamlit~=1.36.0
requests~=2.32.3
python-dotenv~=1.0.1
PyPDF2~=3.0.1
langchain~=0.2.6
fastapi~=0.111.0
uvicorn~=0.30.1
langchain-ai21==0.1.7
langchain-core==0.2.11
langchain-community==0.2.6
packaging~=24.1
numpy~=1.26.4
faiss-cpu~=1.8.0.post1